feat: arreglo de rules en front de lint
he aplicado arreglos de lint y de las rules en la carpeta front
